backend/pyproject.toml
[project]
name = "alloclass-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.46",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "pgvector>=0.4.2",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "umap-learn>=0.5.11",
    "sse-starlette>=3.2.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
